実行内容ごとにファイルを分けた
ファイル名はなんでも良いとのことなので、DB初期化のファイル名は先生の名前を使ってkumackey.goとしました

実行コマンドは下記です
go run *.go

go run main.go だとmain.goファイルしか見ないので実行できないとのことです
